requirements.txt
azure-identity>=1.19.0
azure-keyvault-secrets>=4.9.0
msgraph-sdk>=1.5.0
pydantic>=2.10.0
pyyaml>=6.0.2
